# docs/02_api/openapi.yaml
openapi: 3.0.3
info:
  title: Aviation CAMO & MRO API
  version: 0.2.37
  description: 'Contract baseline for multi-tenant CAMO + MRO + Logistics platform.
    Tenant isolation is implemented via PostgreSQL schema-per-tenant.

    '
servers:
- url: http://localhost:8000
  description: Local dev
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'OIDC access token (Keycloak). Tenant context is resolved from
        token claims.

        '
...
security:
- bearerAuth: []
paths:
...
  /v1/maintenance-events:
    get:
      tags: [Maintenance Events]
      summary: List maintenance events (tenant-isolated)
      parameters:
      - name: aircraft_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaintenanceEventOut'
    post:
      tags: [Maintenance Events]
      summary: Create maintenance event (owner or MRO_EDITOR)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceEventCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceEventOut'
  /v1/maintenance-events/{event_id}:
    delete:
      tags: [Maintenance Events]
      summary: Delete maintenance event (tenant-scoped)
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
  /v1/aircraft/{aircraft_id}/maintenance-events:
    get:
      tags: [Aircraft]
      summary: List maintenance events (owner or assigned MRO)
      parameters:
      - name: aircraft_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AircraftMaintenanceEventOut'
    post:
      tags: [Aircraft]
      summary: Create maintenance event (owner only)
      parameters:
      - name: aircraft_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AircraftMaintenanceEventCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AircraftMaintenanceEventOut'
  /v1/aircraft/{aircraft_id}/maintenance-events/{event_id}:
    put:
      tags: [Aircraft]
      summary: Update maintenance event (owner or assigned MRO)
      parameters:
      - name: aircraft_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: event_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AircraftMaintenanceEventUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AircraftMaintenanceEventOut'
    delete:
      tags: [Aircraft]
      summary: Delete maintenance event (owner or assigned MRO)
      parameters:
      - name: aircraft_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: event_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
...